Move fish set-up to config
I'm not sure that setting *PATH as a global in a set-up script ever
worked, because global variables are not persisted.

Setting *PATH in the config file runs the risk of generating duplicate
values in child shells, but I think these are pretty uncommon. It's also
not necessary, and technically inefficient, to set other variables,
EDITOR etc., in the config file, which is evaluated each time we start
fish. I'm happy to accept these trade-offs for now given they remove the
need for a separate set-up script.