Use LSP tag function when available
This allows us to use a consistent keyboard shortcut to go to a symbol
definition, which invokes the LSP when available and falls back to ctags
when not.

g<C-]> will list multiple definitions when present, and jump to the only
definition otherwise. Remap <C-]> to this.

A next step might be to configure Telescope to search LSP symbols or
tags in the same way.