ctags improvements
Previously, we set the tagfunc to the LSP tagfunc when available. We
also overrode the default tag mapping to use tjump, which will jump to a
single matching tag, or show a list if there are multiple. This allowed
us to use the same mapping to go to definition across projects using
ctags or LSP.

We found that the LSP tagfunc would do a symbol search rather than a
go to definition when called with tjump [1], which meant that we were
seeing irrelevant matches. Switch the solution around, and remap the LSP
go to definition mapping to tjump instead when the filetype is ruby.

Also, switch to universal-ctags, which is a maintained implementation.
It does not require the namespaced classes regex workaround in our
existing ctags file, so remove that.

I was interested to see if we could improve the experience around
namespaced symbols generally, for example by enabling go to definition
with the cursor either side of the separator(s). I don't think this is
possible, because it would require 1) treating the namespaced symbol
as a single word (doable via iskeyword, but with side effects) and
2) a ctags parser that creates namespaced tag entries for symbols
irrespective of how they are defined (namespaced or nested). I think
what we have now (go to definition requires the cursor on the class to
resolve) is the best we're going to get with ctags.

Finally, include ? in the word characters list, so that we can search
for predicate method tags.

[1] https://github.com/neovim/neovim/blob/eae6727325111e596b49bb04337a467e8833397c/runtime/lua/vim/lsp/tagfunc.lua#L76