Set tsserver root_dir to cwd
The default LSP config for tsserver will search upwards from the current
buffer for the first directory with a tsconfig file, and start a server
with that directory as the root. tsserver itself will also search
upwards for a tsconfig file.

In a monorepo environment, with nested tsconfig files, opening a buffer
for a file in a nested package will start a server with that package as
its root. This will leave us unable to detect references outside the
scope of that package. The client for that buffer will be reused for all
files with the same root, regardless of whether we open files (and start
another server) within a parent root.

Switching root_dir to the current working directory enables us to
control the scope of our tsserver session based on the directory in
which we start the editor. This is a niche problem, but it's nice to
document how lspconfig works here, and this new configuration feels more
intuitive.