Set ESLint working directory and timeout
null-ls's default behaviour seems to be to run ESLint in the first
directory above the current file that contains a config file. This was
problematic for a specific configuration including typescript-eslint,
which was expecting to find a tsconfig file relative to that working
directory (which was not present). Setting the working directory to the
editor working directory allowed typescript-eslint to resolve a config
file in a parent directory. ESLint could still resolve its nearest
config file, because it works upwards from the directory containing the
file to be linted. I'm not sure if this will be appropriate for all
cases, but it seems like a reasonable default.

5s was too low a timeout for ESLint, so increase it.

References:

- https://typescript-eslint.io/packages/parser/#project
- https://eslint.org/docs/latest/use/configure/configuration-files#using-configuration-files
- https://marketplace.visualstudio.com/items?itemName=dbaeumer.vscode-eslint